fix: move API route tests out of src/pages/ to prevent Astro build failure
Astro treats all files in src/pages/ as page routes, so test files
containing vi.mock() were being executed during the build, causing
'Vitest mocker was not initialized' errors.

Moved 4 test files to src/__tests__/pages/ and updated their import
paths accordingly.
